[#226] refactor : 로딩 공통 컴포넌트로 분리
